combine test_sql_generation and test_sql_generation_queries.
